Add player.rs and move player logic from main.rs to player.rs
